feat: removed build from repo and added smoke test for whole module
